hack+play explore another approach
I need to write more golang, still struggling with fluency. So this is
me doing more, also I have been writing my utils and CLIs, this will
expand my experience to some basic web services as well.

Also, a better architecture for lcc.live is to be an indirect proxy
rather then a direct proxy. lcc.live should just fetch the images, then
serve those images directly rather then always proxy'ing them.

That will ensure that we are in full control of the percieved
performance of the site, rather then being beholding to all the slow
udot mirrors.

This also opens the door to do other cool things. I have another
experiment that converts these images to video streams, which would
benefit from this re-architecture if i ever get around to it.

Additionally, It's unclear to me that remix was providing much value. It
was fun to play with but ultimately, because of how infrequently I work
on this project, it's dependency instability was a liability for me.